A working version in python
Shell has too many idiosyncracies, and while it seems like the way pv is
run from the shell script is working, if you actually run that command
in a shell it will not start converting until the file is fully piped
through to ffmpeg, rendering the progress bar worthless. How it ever
worked in the first place is a mystery left to our descendents.