Fix #6 - Check for existing files that are being processed.
A start at avoiding collisions from simultaneous (or maybe parallel?) runs.

If a file exists with the name it should have after/during converting,
this tests if the file is a complete conversion. Right now, the only
action taken is to print whether or not the file is verified to be a
complete-copied-transcode. This is because the file might be a
work-in-progress by another process (or potentially this one if
parallel threads are implemented).

In addition:
1. `process_vidlist` calculates and returns the total size saved over
   the course of compressing all of the input files. More information
   about file sizes are printed during the full run as well.
2. Flags are introduced to `process_vidlist` which control whether you
   want the old file deleted, or bad (new) conversions removed. By
   default, operating on a single file (through a CLI argument) does not
   replace the original file to compress. Otherwise, both replacment of
   old files and deletion of (bad) new files are the defaults.
3. The user-defined-limit is now a global variable `NUM_TO_PROCESS`