Types that don't lie
Prior to this, stubs were used to indicate where Enum values would
replace string values (particularly for StreamProperty). This commit
adds the actual classes so the type annotations are sincere. Because of
this, this commit also removes code that performs runtime comparisons
due to lack of type safety.

Now that the StreamProperty class exists and carries with it StreamType
and comparison function information, this commit also removes the
dictionaries that functioned as placeholders for the class and
associated tests.