Show failed tests on bad verification
`compare_stream_property` collects a list of
`StreamPropSingleCompareResult`s in addition to the True/False value
indicating success or failure. Each instance of the new class represents
all comparisons of a given `StreamProperty` for a given `StreamType`,
along with their corresponding type indices. These are then handed back
to the main loop where the results of a bad file comparison can be
printed on the terminal.